react/project code/backend/requirements.txt
# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# === AI / LLM ===
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-openai==0.3.0
langchain-community==0.3.14
langchain-text-splitters==0.3.4
langgraph==0.2.62
google-generativeai==0.8.4

# === Vector Store ===
chromadb==0.5.23

# === Image Processing ===
opencv-python-headless==4.10.0.84
Pillow==11.1.0
pillow-heif==0.21.0

# === Document Export ===
fpdf2==2.8.2
uharfbuzz==0.43.0
python-docx==1.1.2

# === PDF Parsing ===
pymupdf==1.25.3

# === Utilities ===
pydantic==2.10.5
aiofiles==24.1.0
numpy==2.2.2
httpx==0.28.1
